Add module in charge of serizalizing/deserializing
Creation of a module that allows to serialize our
tasks before saving them in the localStorage as
otherwise, they would be saved as blank object
because of closure and private properties.

The module also take care of deserializing, meani-
ng transforming the serialized task as a task obj-
ect again.